requirements.txt
fastapi
gunicorn
uvicorn
sqlalchemy
databases
asyncpg
alembic
mysql-connector-python
python-multipart
pydantic
python-jose
bcrypt
email-validator
passlib
aiosmtplib
jinja2
pytest
pytest-asyncio
httpx
asyncmy
pydantic-settings
greenlet
cryptography
pytest
pytest-asyncio
